ts: Move assertQuery() method to correct file
This one method of the testModelRunner was located in a different file
for some reason, moved for consistency.

Release note: None
